linting work

+++ src/qattr.c
@@ -220,7 +220,7 @@ qattr_list_attr_key_get(const QattrList_t *attr_list, int index) {
 		Q_ERRORFOUND(QERROR_NULL_POINTER_UNEXPECTED);
 		return (QattrKey_t) Q_ERRORCODE_ENUM;
 	}
-	if (index >= qattr_list_index_ok_get(attr_list)) {
+	if (index >= (int) qattr_list_index_ok_get(attr_list)) {
 		Q_ERRORFOUND(QERROR_INDEX_OUTOFRANGE);
 		return (QattrKey_t) Q_ERRORCODE_ENUM;
 	}
